Divide code in main.js into modules (util.js and data.js) and create basic structure of the project
